Reduce verbosity for screen reader users
This commit swaps all the `article` tags to `div` tags. It's because
`article` can make screen reader list all the `article` tag. Then, it
lists the heading of the article as the label or accessible. It's the
same as listing all the heading tags.

So, by using `div` the screen readers can only list headings.

It also swaps the list item inside the main landmark with `div`. All the main
content can be understood without list element.