Remove social media list item from `footer`
This commit makes the social media links less verbose to be pronounced
by screen readers. Also, social media links can be understood without it
being a list.

It's worth to note that I make all the links (`a`) to have `height:
fit-content` to prevent any links from inheriting the height of the
parent element.